agent-backend/requirements.txt
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn>=0.30.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
aiosqlite>=0.20.0
beautifulsoup4>=4.12.0
python-multipart>=0.0.9
